+++ app/src/main/java/org/techconnect/activities/MainActivity.java
@@ -20,9 +20,7 @@ import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.Gravity;
-import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.FrameLayout;
@@ -31,8 +29,6 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import org.centum.techconnect.R;
-import org.techconnect.adapters.CategoryListAdapter;
-import org.techconnect.adapters.SessionCursorAdapter;
 import org.techconnect.analytics.FirebaseEvents;
 import org.techconnect.asynctasks.LogoutAsyncTask;
 import org.techconnect.asynctasks.PostAppFeedbackAsyncTask;
@@ -97,11 +93,6 @@ public class MainActivity extends AppCompatActivity
     private boolean showedLogin = false;
     private boolean userLearnedDrawer = false;
     private Stack<Integer> fragStack = new Stack<>();
-
-    //Adapters for Repair History Fragment Menu
-    private SessionCursorAdapter sessionAdapter;
-    private CategoryListAdapter dateAdapter = new CategoryListAdapter();
-    private CategoryListAdapter deviceAdapter = new CategoryListAdapter();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -395,70 +386,6 @@ public class MainActivity extends AppCompatActivity
         }
     }
 
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.activity_main_toolbar_menu, menu);
-        MenuItem item = menu.findItem(R.id.action_sort);
-        item.setVisible(false);
-        //Initially, will have date be the initial way to sort the sessions
-        item.getSubMenu().findItem(R.id.date_item).setChecked(true);
-
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        if (item.isChecked())
-            item.setChecked(false);
-        else
-            item.setChecked(true);
-
-        switch(item.getItemId()) {
-            case R.id.date_item:
-                Log.d("Repair History","DATE");
-                ((RepairHistoryFragment) FRAGMENTS[FRAGMENT_HISTORY]).setAdapter(dateAdapter);
-                //categoryAdapter.setBaseMap(date_counts);
-                break;
-            case R.id.device_item:
-                Log.d("Repair History","DEVICE");
-                ((RepairHistoryFragment) FRAGMENTS[FRAGMENT_HISTORY]).setAdapter(deviceAdapter);
-                //categoryAdapter.setBaseMap(device_counts);
-                break;
-            case R.id.action_sort:
-                Log.d("Repair History","SORT");
-                break;
-            default:
-                Log.d("Repair History","DEVICE");
-                //categoryAdapter.setBaseMap(null);
-        }
-        return true;
-    }
-
-    public SessionCursorAdapter getSessionAdapter() {
-        return sessionAdapter;
-    }
-
-    public void setSessionAdapter(SessionCursorAdapter sessionAdapter) {
-        this.sessionAdapter = sessionAdapter;
-    }
-
-    public CategoryListAdapter getDeviceAdapter() {
-        return deviceAdapter;
-    }
-
-    public void setDeviceAdapter(CategoryListAdapter deviceAdapter) {
-        this.deviceAdapter = deviceAdapter;
-    }
-
-    public CategoryListAdapter getDateAdapter() {
-        return dateAdapter;
-    }
-
-    public void setDateAdapter(CategoryListAdapter dateAdapter) {
-        this.dateAdapter = dateAdapter;
-    }
 }
